+++ app/lib/Functions/bulkUpdate.js
@@ -1,7 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.bulkUpdate = void 0;
-const lodash_1 = require("lodash");
 const connection_helpers_1 = require("../Helpers/connection_helpers");
 const item_helpers_1 = require("../Helpers/item_helpers");
 async function bulkUpdate(collectionId, items, options) {
@@ -14,42 +13,32 @@ async function bulkUpdate(collectionId, items, options) {
                 throw Error(`WeivData - Item (_id) ID is required for each item when bulk updating ID is missing for one or more item in your array!`);
             }
         }
-        const { suppressAuth, suppressHooks, cleanupAfter, enableOwnerId, consistentRead } = options || { suppressAuth: false, suppressHooks: false, cleanupAfter: false, enableOwnerId: true };
-        const defaultValues = {
-            _updatedDate: new Date()
-        };
+        const { suppressAuth, suppressHooks, cleanupAfter, consistentRead } = options || { suppressAuth: false, suppressHooks: false, cleanupAfter: false };
         const editedItems = items.map((item) => {
             item._id = (0, item_helpers_1.convertStringId)(item._id);
-            item = (0, lodash_1.merge)(defaultValues, item);
-            return item;
+            return {
+                ...item,
+                _updatedDate: new Date()
+            };
         });
-        const query = {
-            _id: { $in: editedItems.map((item) => (0, item_helpers_1.convertStringId)(item._id)) },
-        };
-        const updateObjects = editedItems.map((item) => ({
-            $set: item.updatedFields,
-        }));
+        const bulkOperations = editedItems.map((item) => {
+            return {
+                updateOne: {
+                    filter: { _id: item._id },
+                    update: { $set: item }
+                }
+            };
+        });
+        console.log(editedItems, bulkOperations);
         const { collection, cleanup } = await (0, connection_helpers_1.connectionHandler)(collectionId, suppressAuth);
-        let succeed = true;
-        let updated = 0;
-        for (let i = 0; i < editedItems.length; i += 50) {
-            const updateBatch = updateObjects.slice(i, i + 50);
-            const { modifiedCount, acknowledged } = await collection.updateMany(query, updateBatch, { readConcern: consistentRead === true ? "majority" : "local" });
-            succeed = acknowledged;
-            updated = updated + modifiedCount;
-        }
+        const { matchedCount } = await collection.bulkWrite(bulkOperations, { readConcern: consistentRead === true ? "majority" : "local" });
         if (cleanupAfter === true) {
             await cleanup();
         }
-        if (succeed === true) {
-            return {
-                updated,
-                updatedItemIds: editedItems.map((item) => item._id)
-            };
-        }
-        else {
-            throw Error(`WeivData - Error when updating items using bulkUpdate, acknowledged: ${succeed}, updated: ${updated}`);
-        }
+        return {
+            updated: matchedCount,
+            updatedItems: editedItems
+        };
     }
     catch (err) {
         throw Error(`WeivData - Error when updating items using bulkUpdate: ${err}`);